+++ presentation-core/src/main/java/tachiyomi/presentation/core/components/material/PullRefresh.kt
@@ -3,8 +3,6 @@ package tachiyomi.presentation.core.components.material
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.pulltorefresh.PullToRefreshDefaults
 import androidx.compose.material3.pulltorefresh.pullToRefresh
 import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
 import androidx.compose.runtime.Composable
@@ -40,14 +38,12 @@ fun PullRefresh(
     ) {
         content()
 
-        PullToRefreshDefaults.Indicator(
+        LiquidPullRefreshIndicator(
             modifier = Modifier
                 .align(Alignment.TopCenter)
                 .padding(indicatorPadding),
             isRefreshing = refreshing,
             state = state,
-            containerColor = MaterialTheme.colorScheme.surfaceVariant,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
         )
     }
 }